requirements.txt
flask
flask-cors
python-dotenv
requests
gspread
google-auth
pandas
pymupdf
gunicorn
psycopg2-binary
